listin

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -8,7 +8,7 @@
 
 size_t print_listint(const listint_t *h)
 {
-	float a = 0;
+	int a = 0;
 
 	while (h)
 	{
